pyproject.toml
[tool.poetry]
name = "webprotocols"
version = "0.1.0"
description = ""
authors = ["Yaroslav Martynenko <stikblacklabel@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = {extras = ["all"], version = "^0.77.0"}
uvicorn = "^0.17.6"
strawberry-graphql = {extras = ["debug-server"], version = "^0.114.0"}
fastapi-jsonrpc = "^2.2.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
